+++ py/bitbox02/bitbox02/bitbox02/bitbox02.py
@@ -788,7 +788,13 @@ class BitBox02(BitBoxCommonAPI):
         )
         return self._eth_msg_query(request, expected_response="pub").pub.pub
 
-    def eth_sign(self, transaction: bytes, keypath: Sequence[int], chain_id: int = 1) -> bytes:
+    def eth_sign(
+        self,
+        transaction: bytes,
+        keypath: Sequence[int],
+        address_case: eth.ETHAddressCase.ValueType = eth.ETH_ADDRESS_CASE_MIXED,
+        chain_id: int = 1,
+    ) -> bytes:
         """
         transaction should be given as a full rlp encoded eth transaction.
         """
@@ -853,6 +859,7 @@ class BitBox02(BitBoxCommonAPI):
                     recipient=recipient,
                     value=value,
                     data=data,
+                    address_case=address_case,
                 )
             )
             return handle_antiklepto(request)
@@ -871,6 +878,7 @@ class BitBox02(BitBoxCommonAPI):
                 recipient=recipient,
                 value=value,
                 data=data,
+                address_case=address_case,
             )
         )